feat: Replace legacy delegate system with stateless architecture
- Remove all legacy delegate-based session management files
- Implement StatelessSessionManager using StatelessAgent + StatelessAILoop
- Add built-in WebSocket streaming without patches or workarounds
- Clean up main.py to use stateless architecture exclusively
- Fix AI response streaming that was broken in legacy system

Major architectural improvement:
- No more fragile delegate/bridge complexity
- Direct agent-to-websocket streaming
- Cleaner, more maintainable codebase
- Confirmed working: agent switching, AI responses, streaming
