+++ sessions.json
@@ -1,7 +1,7 @@
 {
   "sessions": [
-    "3f29ea10-1024-4c48-bc38-f0227823590d",
-    "50060226-5e30-4957-9bd0-d49d8a84ce7e"
+    "b70e5c5a-0334-4f76-a730-d90fb05c737b",
+    "44a53c3a-df10-44a3-a3de-c4328b8fe864"
   ],
-  "timestamp": "2025-05-28T19:25:29.059816"
+  "timestamp": "2025-05-28T19:45:37.448474"
 }
